Add support for the commit message
There's a new commit function added to the vfs trait, and a new
write fuction that receive the sync level, and returns the sync
level. The default implementation of the latter defers to the
exisiting write function.

nfs_handler was updated to make use of these new functions.

I may completely fork at some point and lose support for the original
write function, but I haven't committed to that yet.